Added dmz project.
The combination of package objects and pervasive use of
type aliases is a lethal combination for scala, and the bugs
tend to manifest as mysteriously missing symbols with no error
ever issued or even hinted at. I cooked up this structure
so that the more interesting developments can be built on this
stable core of type aliases and some minimal machinery.