Breaking free of the default namespace.
Now building everything with -Yno-predef and -Yno-imports.
Can't use -Yno-imports in the stock repl, it can't handle it.
Working toward reclaiming all the simple names.

This commit is a big mess, it's carved off of a larger one
and it didn't go very smoothly. We'll be going back to the
larger one and agree to forget about this one.