Improving terminology.
Let's call the set of values for which a map is defined
its "domain".

Really we should think of maps as translucent functions.
They reveal for which values they are well-defined (either
via a predicate or an explicit set of keys) and new maps
can be derived from them. This translucency is what
PartialFunctions were supposed to accomplish in scala.