Fiddling with sets and maps.
Found a super subtle scaryish bug. You really have to be careful
what you name the thing in your implicit classes. Mostly I use x
but in one case I used "pf" and that was a bit too likely to be
the same as an actual field. Or at least, what looked like an
actual field, but wasn't, but compiled anyway because of the
implicit conversion - providing a different partial function than
the one intended as a consequence. Sketchville.