Trim the wrong kind of innovation.
Slipping too far into implementing things which I shouldn't
be implementing. My avenue of innovation here is safety,
coherency, and general usability, not in the low level
implementations, which in any case are the easiest things
to improve later.

So this represents the first in a hopefully short series of
commits which should leave us with thin wrappers around existing
collections-like classes from scala and java.